fix: make fail use correct attribute name
The prior change incorrectly made `_fail` look for the `env._failures`
attribute, when it is actually called `env.failures`

This went unnoticed because all the tests mock out the fail call
(if they didn't, then tests checking for expected failures would fail).

To work around this, add a test that calls the real failure function
to exercise the real code path. The test clears the list of failures
prior to the test finishing to prevent the expected failure from
failing the entire test.